Ver. 1.7

+++ src/br/com/partypoker/dao/IDAOGeneric.java
@@ -11,7 +11,7 @@ public interface IDAOGeneric<E extends Entidade> {
 	
 	public void create(E e) throws DaoException;
 
-	public E search(Class<E> classe, int id) throws DaoException;
+	public E search(Class<E> classe, Long id) throws DaoException;
 
 	public void disable(E e) throws DaoException;
